backend/requirements.txt
fastapi
uvicorn
python-multipart
langchain
langchain-openai
langchain-community
faiss-cpu
pypdf
python-dotenv
tiktoken
pandas
openpyxl
numpy<2
langchain-core
langchain-experimental
langchain-text-splitters
matplotlib
seaborn
tabulate
